Import from mapdata service for Hamburg
