+++ run-preheating.sh
@@ -28,7 +28,7 @@ if [ $ret -ne 0 ]; then
 fi
 
 # Load the construction sites
-poetry run python backend/manage.py import_constructions_overpass ${LOCATION}
+poetry run python backend/manage.py import_constructions ${LOCATION}
 
 # Check if previous command failed. If it did, exit
 ret=$?
